+++ HISTORY.rst
@@ -8,7 +8,8 @@ History
 
 * Add support for multisite setups
 * Drop django CMS 3.0 compatibility
-* Add Django 1.9 support
+* Add Django 1.9, 1.10 support
+* Add django CMS 3.4 support
 
 0.1.2 (2015-12-22)
 ++++++++++++++++++